+++ src/cli.rs
@@ -329,7 +329,8 @@ pub fn build_cli() -> ArgMatches {
                 .takes_value(true)
                 .multiple_occurrences(false)
                 .require_equals(false)
-                .validator(filename_pattern_validator)
+                .required(false)
+                .validator(crate::shared::file_rename_pattern_validate)
                 .hide(false).help_heading(operations_name)
                 .display_order(1)
         )
@@ -350,18 +351,5 @@ fn genre_number_validator(input: &str) -> Result<(), String> {
         Err(_) => Err(String::from(
             "Unable to parse the input provided to --track-genre-number.",
         )),
-    }
-}
-
-/// Checks that the new filename pattern results in a unique file
-fn filename_pattern_validator(pattern: &str) -> Result<(), String> {
-    if !pattern.contains("%tn")
-        && !pattern.contains("%tt")
-        && !pattern.contains("%track-number")
-        && !pattern.contains("%track-title")
-    {
-        Err(format!("Pattern \"{}\" would not yield unique file names. Pattern must contain track number and/or track name. Cannot continue.", pattern))
-    } else {
-        Ok(())
     }
 }