+++ src/default_values.rs
@@ -116,30 +116,8 @@ impl DefaultValues {
             log::debug!("Loaded config: {:?}", &config);
         }
 
-        // Verify the rename_file pattern
-        let mut pattern = "".to_string();
-        if let Some(rnp) = cli_args.value_of("rename-file") {
-            pattern = rnp.to_string();
-        }
-
-        if cli_args.is_present("rename-file") {
-            pattern = cli_args
-                .value_of("rename-file")
-                .unwrap_or_default()
-                .to_string();
-        };
-
-        if !pattern.contains("%tn")
-            && !pattern.contains("%tt")
-            && !pattern.contains("%track-number")
-            && !pattern.contains("%track-title")
-        {
-            return Err(format!("Pattern \"{}\" would not yield unique file names. Pattern must contain track number and/or track name. Cannot continue.", pattern).into());
-        } else {
-            config.rename_file = Some(pattern);
-        }
-
         // Collate config file flags and CLI flags and output the right config
+        check_for_rename(&mut config, cli_args)?; // Check if we have a file rename
         config.quiet = Some(quiet(&config, cli_args));
         config.stop_on_error = Some(stop_on_error(&config, cli_args));
         config.print_summary = Some(print_summary(&config, cli_args));
@@ -301,4 +279,25 @@ fn dry_run(defaults: &DefaultValues, args: &clap::ArgMatches) -> bool {
 
     // return the value
     return_value
+}
+
+/// Checks the loaded config if there is a `file_rename` present, and validates it.
+/// Also checks the CLI for a rename-file and overrides any previous config entries if it is present.
+/// Returns OK if everything went well. Returns an error if the `file_rename` is invalid.
+fn check_for_rename(
+    config: &mut DefaultValues,
+    args: &clap::ArgMatches,
+) -> Result<(), Box<dyn Error>> {
+    // Check if anything came from the config file and validate it
+    if let Some(rnp) = &config.rename_file {
+        crate::shared::file_rename_pattern_validate(rnp)?;
+    }
+
+    // Even if we have something from the config file, CLI takes presedence
+    if args.is_present("rename-file") {
+        config.rename_file = Some(args.value_of("rename-file").unwrap_or_default().to_string());
+    };
+
+    // Return safely
+    Ok(())
 }